droneberry/push_notifications/requirements.txt
netifaces==0.10.6
python-pushsafer==0.2